move Number to I18n
